.github/actions/bump-my-version/requirements.txt
requests==2.32.3 
GitPython==3.1.44
urllib3==2.4.0